requirements.txt
angles==2.0
ftfy==6.1.1
gdown==4.6.0
matplotlib==3.6.0
numpy==1.23.1
pandas==1.5.0
Pillow==9.5.0
pyecharts==2.0.3
PyYAML==6.0
regex==2022.10.31
scikit_learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
snapshot_selenium==0.0.2
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.1
